+++ src/ui/hooks/index.ts
@@ -1,0 +1,3 @@
+import useSettings from '@/ui/hooks/useSettings'
+
+export { useSettings }
